third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/module/referrer-origin-when-cross-origin.sub-expected.txt

This is a testharness.js-based test.
PASS Importing a same-origin top-level script with the origin-when-cross-origin policy.
PASS Importing a remote-origin top-level script with the origin-when-cross-origin policy.
PASS Importing a same-origin descendant script from a same-origin top-level script with the origin-when-cross-origin policy.
PASS Importing a remote-origin descendant script from a same-origin top-level script with the origin-when-cross-origin policy.
FAIL Importing a remote-origin descendant script from a remote-origin top-level script with the origin-when-cross-origin policy. assert_equals: Referrer should be sent for the remote-origin descendant script. expected "http://www1.web-platform.test:8001/" but got "http://www1.web-platform.test:8001/html/semantics/scripting-1/the-script-element/module/resources/import-referrer-checker.sub.js?name=remote_remote"
Harness: the test ran to completion.
